test: migrate regression test to unit tests
