+++ engine/src/test/java/io/camunda/zeebe/engine/processing/streamprocessor/ProcessingScheduleServiceIntegrationTest.java
@@ -11,17 +11,13 @@ import static io.camunda.zeebe.engine.util.RecordToWrite.command;
 import static io.camunda.zeebe.engine.util.RecordToWrite.event;
 import static io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent.ACTIVATE_ELEMENT;
 import static io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent.ELEMENT_ACTIVATING;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
-import io.camunda.zeebe.engine.Loggers;
 import io.camunda.zeebe.engine.api.EmptyProcessingResult;
 import io.camunda.zeebe.engine.api.ProcessingResult;
 import io.camunda.zeebe.engine.api.ProcessingResultBuilder;
@@ -36,17 +32,12 @@ import io.camunda.zeebe.engine.api.records.RecordBatch;
 import io.camunda.zeebe.engine.util.Records;
 import io.camunda.zeebe.engine.util.StreamPlatform;
 import io.camunda.zeebe.engine.util.StreamPlatformExtension;
-import io.camunda.zeebe.logstreams.log.LoggedEvent;
 import io.camunda.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceRecord;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.scheduler.clock.ControlledActorClock;
-import io.camunda.zeebe.scheduler.future.CompletableActorFuture;
-import io.camunda.zeebe.test.util.junit.RegressionTest;
 import java.time.Duration;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicInteger;
-import org.awaitility.Awaitility;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -202,73 +193,6 @@ public class ProcessingScheduleServiceIntegrationTest {
     // then
     verify(dummyProcessorSpy, TIMEOUT.times(5))
         .process(Mockito.argThat(record -> record.getKey() == 1), any());
-  }
-
-  @RegressionTest("https://github.com/camunda/zeebe/issues/10240")
-  void shouldPreserveOrderingOfWritesEvenWithRetries() {
-    // given
-    final var dummyProcessorSpy = spy(dummyProcessor);
-    final var syncLogStream = spy(streamPlatform.getLogStream());
-    final var logStream = spy(syncLogStream.getAsyncLogStream());
-    final var batchWriter = spy(syncLogStream.newLogStreamBatchWriter());
-
-    when(syncLogStream.getAsyncLogStream()).thenReturn(logStream);
-    doReturn(CompletableActorFuture.completed(batchWriter))
-        .when(logStream)
-        .newLogStreamBatchWriter();
-    streamPlatform
-        .withRecordProcessors(List.of(dummyProcessorSpy))
-        .buildStreamProcessor(syncLogStream, true);
-
-    // when - in order to make sure we would interleave tasks without the fix for #10240, we need to
-    // make sure we retry at least twice, such that the second task can be executed in between both
-    // invocations. ensure both tasks have an expiry far away enough such that they expire on
-    // different ticks, as tasks expiring on the same tick will be submitted in a non-deterministic
-    // order
-    final var counter = new AtomicInteger(0);
-    when(batchWriter.tryWrite())
-        .then(
-            i -> {
-              final var invocationCount = counter.incrementAndGet();
-              // wait a sufficiently high enough invocation count to ensure the second timer is
-              // expired, gets scheduled, and then the executions are interleaved. this is quite
-              // hard to do in a deterministic controlled way because of the way our timers are
-              // scheduled
-              if (invocationCount < 5000) {
-                return -1L;
-              }
-
-              Loggers.PROCESS_PROCESSOR_LOGGER.debug("Calling real method");
-              return i.callRealMethod();
-            });
-
-    // we need to schedule first the longer delayed task, otherwise we might get race conditions
-    // with the scheduling and clock adjustment
-    dummyProcessorSpy.scheduleService.runDelayed(
-        Duration.ofMinutes(1),
-        builder -> {
-          Loggers.PROCESS_PROCESSOR_LOGGER.debug("Running second timer");
-          builder.appendCommandRecord(2, ACTIVATE_ELEMENT, RECORD);
-          return builder.build();
-        });
-    dummyProcessorSpy.scheduleService.runDelayed(
-        Duration.ZERO,
-        builder -> {
-          Loggers.PROCESS_PROCESSOR_LOGGER.debug("Running first timer");
-          // force trigger second task
-          clock.addTime(Duration.ofMinutes(1));
-          builder.appendCommandRecord(1, ACTIVATE_ELEMENT, RECORD);
-          return builder.build();
-        });
-
-    // then
-    Awaitility.await("until both records are written to the stream")
-        .atMost(Duration.ofSeconds(10))
-        .untilAsserted(() -> assertThat(streamPlatform.events()).hasSize(2));
-    assertThat(streamPlatform.events())
-        .as("records were written in order of submitted tasks")
-        .extracting(LoggedEvent::getKey)
-        .containsExactly(1L, 2L);
   }
 
   private static final class DummyTask implements Task {
